Fix only one remand reason gets sent
