+++ client/app/queue/SelectRemandReasonsView.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import _ from 'lodash';
@@ -8,7 +9,9 @@ import COPY from '../../COPY.json';
 
 import decisionViewBase from './components/DecisionViewBase';
 import IssueRemandReasonsOptions from './components/IssueRemandReasonsOptions';
-
+import {
+  editStagedAppeal
+} from './QueueActions';
 import {
   fullWidth,
   VACOLS_DISPOSITIONS,
@@ -41,11 +44,12 @@ class SelectRemandReasonsView extends React.Component {
   goToPrevStep = () => _.each(this.state.renderedChildren, (child) => child.updateStoreIssue());
 
   goToNextStep = () => {
-    const { issues } = this.props;
+    const { issues, appealId, appeal, amaDecisionIssues } = this.props;
     const {
       issuesRendered,
       renderedChildren
     } = this.state;
+    const useDecisionIssues = !appeal.isLegacyAppeal && (amaDecisionIssues || appeal.decisionIssues);
 
     if (issuesRendered < issues.length) {
       this.setState({ issuesRendered: Math.min(issuesRendered + 1, issues.length) });
@@ -53,7 +57,10 @@ class SelectRemandReasonsView extends React.Component {
       return false;
     }
 
-    _.each(renderedChildren, (child) => child.updateStoreIssue());
+    const updatedIssues = _.map(renderedChildren, (child) => child.updateStoreIssue());
+    const attributes = useDecisionIssues ? { decisionIssues: updatedIssues } : { issues: updatedIssues };
+
+    this.props.editStagedAppeal(appealId, attributes);
 
     return true;
   }
@@ -116,8 +123,13 @@ const mapStateToProps = (state, ownProps) => {
     issues: _.filter(issues, (issue) => [
       VACOLS_DISPOSITIONS.REMANDED, ISSUE_DISPOSITIONS.REMANDED
     ].includes(issue.disposition)),
-    ..._.pick(state.ui, 'userRole')
+    ..._.pick(state.ui, 'userRole'),
+    amaDecisionIssues: state.ui.featureToggles.ama_decision_issues
   };
 };
 
-export default connect(mapStateToProps)(decisionViewBase(SelectRemandReasonsView));
+const mapDispatchToProps = (dispatch) => bindActionCreators({
+  editStagedAppeal
+}, dispatch);
+
+export default connect(mapStateToProps, mapDispatchToProps)(decisionViewBase(SelectRemandReasonsView));
